feat: prepare pages transltations (with meta)
