--- messages/en.json
@@ -1,13 +1,0 @@
-{
-  "layout": {
-    "nav": {
-      "story": "Story",
-      "dashboard": "Dashboard",
-      "act": "To act",
-      "about": "About"
-    }
-  },
-  "story": {
-    "title": "The salmon you eat, sell, produce..."
-  }
-}
