+++ messages/en/layout.json
@@ -1,0 +1,15 @@
+{
+  "layout": {
+    "nav": {
+      "story": "Story",
+      "dashboard": "Dashboard",
+      "act": "To act",
+      "about": "About"
+    }
+  },
+  "siteConfig": {
+    "title": "Pinkbombs",
+    "description": "Analyzing the impact of Salmon on the Environment, Ecosystem, Health and Communities.",
+    "url": "https://pinkbombs-app.vercel.app"
+  }
+}
