+++ messages/en/pages/act.json
@@ -1,0 +1,5 @@
+{
+  "act": {
+    "title": ""
+  }
+}
