+++ messages/en/pages/story.json
@@ -1,0 +1,5 @@
+{
+  "story": {
+    "title": "The salmon you eat, sell, produce..."
+  }
+}
