--- messages/fr.json
@@ -1,13 +1,0 @@
-{
-  "layout": {
-    "nav": {
-      "story": "Story",
-      "dashboard": "Dashboard",
-      "act": "Agir",
-      "about": "À propos"
-    }
-  },
-  "story": {
-    "title": "Le saumon que vous mangez, vendez, produisez..."
-  }
-}
