+++ messages/fr/components.json
@@ -1,0 +1,3 @@
+{
+  "components": {}
+}
