+++ messages/fr/pages/about.json
@@ -1,0 +1,5 @@
+{
+  "about": {
+    "title": ""
+  }
+}
