+++ messages/fr/pages/dashboard.json
@@ -1,0 +1,5 @@
+{
+  "dashboard": {
+    "title": "Dashboard"
+  }
+}
