+++ messages/fr/pages/story.json
@@ -1,0 +1,5 @@
+{
+  "story": {
+    "title": "Le saumon que vous mangez, vendez, produisez..."
+  }
+}
