+++ src/app/[locale]/about/about.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import * as React from "react";
+import "@/lib/env";
+
+import IntroBlock from "@/components/IntroBlock";
+
+const About = () => {
+  return (
+    <>
+      <IntroBlock title="À propos" />
+    </>
+  );
+};
+
+export default About;
