+++ src/app/[locale]/to-act/act.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import * as React from "react";
+import "@/lib/env";
+
+import IntroBlock from "@/components/IntroBlock";
+
+const Act = () => {
+  return (
+    <>
+      <IntroBlock title="Agir" />
+    </>
+  );
+};
+
+export default Act;
